refector : api 및 이미지 등 리펙토링
